fix types

+++ osmnx/truncate.py
@@ -16,7 +16,7 @@ from . import utils_graph
 def truncate_graph_dist(
     G: nx.MultiDiGraph,
     source_node: int,
-    max_dist: int = 1000,
+    max_dist: float = 1000,
     weight: str = "length",
     retain_all: bool = False,
 ) -> nx.MultiDiGraph:
